feat: implement weight-tracking feature + reviewer-ready README
Full implementation of the weight & BMI tracking feature (spec T-00→T-29):
- Next.js 14 App Router + TypeScript + Tailwind
- SQLite + Drizzle ORM (schema, migrations, singleton connection)
- Pure domain layer: calculateBMI + WHO classification
- Zod schemas shared between API routes and client forms
- Repositories with per-user isolation
- API routes: profile (GET/PUT), weight (GET/POST/PUT/DELETE)
- UI: profile form, weight entry form with live BMI preview,
  history table, accessible delete dialog, BMI category badges
- Tests: 22 passing (9 unit + 13 property-based with fast-check)
- Seed script: 1000-row performance check (history query ~1ms)

Docs:
- README rewritten with quick-start and validation guide for reviewers
- progress-log.md documenting every task and decision
- Include packaged Power (Bonus 2) as regular files (removed nested .git)
- gitignore build cache (tsbuildinfo, next-env.d.ts)

Kiro University Challenge 2026 — Lessons 1-7 + Bonus 1 & 2

+++ src/domain/bmi.property.test.ts
@@ -23,10 +23,10 @@ import { calculateBMI, type BMICategory } from './bmi';
 // ---------------------------------------------------------------------------
 
 /** Peso válido según RF-02: (0, 700] kg */
-const validWeight = () => fc.float({ min: 0.1, max: 700, noNaN: true });
+const validWeight = () => fc.double({ min: 0.1, max: 700, noNaN: true });
 
 /** Altura válida según el modelo: (0, 300] cm — rango humano razonable */
-const validHeight = () => fc.float({ min: 1, max: 300, noNaN: true });
+const validHeight = () => fc.double({ min: 1, max: 300, noNaN: true });
 
 /** Par (weightKg, heightCm) siempre válido */
 const validPair = () =>
@@ -118,8 +118,8 @@ describe('Invariante: monotonía respecto al peso', () => {
     fc.assert(
       fc.property(
         validHeight(),
-        fc.float({ min: 0.1, max: 349.9, noNaN: true }),
-        fc.float({ min: 350, max: 700, noNaN: true }),
+        fc.double({ min: 0.1, max: 349.9, noNaN: true }),
+        fc.double({ min: 350, max: 700, noNaN: true }),
         (heightCm, lightWeight, heavyWeight) => {
           const { value: bmiLight } = calculateBMI(lightWeight, heightCm);
           const { value: bmiHeavy } = calculateBMI(heavyWeight, heightCm);
@@ -139,8 +139,8 @@ describe('Invariante: monotonía respecto a la altura', () => {
     fc.assert(
       fc.property(
         validWeight(),
-        fc.float({ min: 1, max: 149, noNaN: true }),
-        fc.float({ min: 150, max: 300, noNaN: true }),
+        fc.double({ min: 1, max: 149, noNaN: true }),
+        fc.double({ min: 150, max: 300, noNaN: true }),
         (weightKg, shortHeight, tallHeight) => {
           const { value: bmiShort } = calculateBMI(weightKg, shortHeight);
           const { value: bmiTall } = calculateBMI(weightKg, tallHeight);
@@ -169,8 +169,8 @@ describe('Invariante: clasificación WHO (RF-04)', () => {
     // Genera pares cuyo IMC de referencia caiga en < 18.5
     fc.assert(
       fc.property(
-        fc.float({ min: 1, max: 300, noNaN: true }),  // heightCm
-        fc.float({ min: 0.001, max: 1, noNaN: true }), // factor < 1 para forzar bajo IMC
+        fc.double({ min: 1, max: 300, noNaN: true }),  // heightCm
+        fc.double({ min: 0.001, max: 1, noNaN: true }), // factor < 1 para forzar bajo IMC
         (heightCm, factor) => {
           // weight tal que IMC exacto = 18.4 * factor  → siempre < 18.5
           const targetBMI = 18.49 * factor; // estrictamente < 18.5
@@ -188,8 +188,8 @@ describe('Invariante: clasificación WHO (RF-04)', () => {
   it('todos los valores ≥ 30.0 se clasifican como obese', () => {
     fc.assert(
       fc.property(
-        fc.float({ min: 1, max: 300, noNaN: true }),
-        fc.float({ min: 1, max: 10, noNaN: true }), // multiplicador ≥ 1 para empujar IMC alto
+        fc.double({ min: 1, max: 300, noNaN: true }),
+        fc.double({ min: 1, max: 10, noNaN: true }), // multiplicador ≥ 1 para empujar IMC alto
         (heightCm, multiplier) => {
           const targetBMI = 30.0 * multiplier; // siempre ≥ 30
           const heightM = heightCm / 100;
@@ -216,7 +216,7 @@ describe('Invariante: escala cuadrática de la altura', () => {
     fc.assert(
       fc.property(
         validPair(),
-        fc.float({ min: 0.5, max: 2.0, noNaN: true }), // factor de escala k
+        fc.double({ min: 0.5, max: 2.0, noNaN: true }), // factor de escala k
         ({ weightKg, heightCm }, k) => {
           const scaledWeight = weightKg * k * k;
           const scaledHeight = heightCm * k;
@@ -242,7 +242,7 @@ describe('Invariante: rechazo de entradas inválidas', () => {
       fc.property(
         fc.oneof(
           fc.constant(0),
-          fc.float({ min: -1e6, max: -Number.EPSILON, noNaN: true }),
+          fc.double({ min: -1e6, max: -Number.EPSILON, noNaN: true }),
         ),
         validHeight(),
         (badWeight, heightCm) => {
@@ -258,7 +258,7 @@ describe('Invariante: rechazo de entradas inválidas', () => {
         validWeight(),
         fc.oneof(
           fc.constant(0),
-          fc.float({ min: -1e6, max: -Number.EPSILON, noNaN: true }),
+          fc.double({ min: -1e6, max: -Number.EPSILON, noNaN: true }),
         ),
         (weightKg, badHeight) => {
           expect(() => calculateBMI(weightKg, badHeight)).toThrow(RangeError);
